+++ src/api/clientDocumentType/clientDocumentType.service.ts
@@ -17,6 +17,7 @@ import { getSortingParams } from "../../components/sorting";
 import { CompanyModel } from "../company";
 import { ClientDocumentCategoryModel } from "../clientDocumentCategory";
 import { getFilters } from "../../components/filters";
+import { clientDocumentService } from "../clientDocument";
 
 class ClientDocumentTypeService {
   async createBulkClientDocumentType(props: CreateBulkClientDocumentTypeProps) {
@@ -111,6 +112,20 @@ class ClientDocumentTypeService {
   async deleteClientDocumentType(props: DeleteClientDocumentTypeProps) {
     // Props
     const { id, company } = props;
+
+    // Check if this document type is used by any document
+    const clientDocument = await clientDocumentService.getClientDocumentByType({
+      type: id,
+      company,
+    });
+
+    // If used, then don't allow the user to delete this
+    if (clientDocument && clientDocument.length > 0) {
+      throw new CustomError(
+        409,
+        ClientDocumentTypeErrorCode.DOCUMENT_TYPE_IN_USE
+      );
+    }
 
     // Find and delete the clientDocumentType by id and company
     const clientDocumentType = await ClientDocumentTypeModel.destroy({
